services/frontend/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
    "streamlit>=1.40.2",
    "azure-ai-projects>=1.0.0b2",
    "azure-ai-inference>=1.0.0b6",
    "azure-identity>=1.19.0",
    "azure-search-documents",
    "pandas",
    "python-dotenv",
    "opentelemetry-api",
    "opentelemetry-sdk>=1.28.2",
    "azure-monitor-opentelemetry>=1.6.4",
    "azure-ai-evaluation>=1.0.1",
    "pyyaml>=6.0.2",
]
